pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srl-rl"
version = "0.2.0"
description = "Simple Reinforcement Learning Library — continuous action spaces, multi-modal, config-driven"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "Tran Dang An", email = "bigkatoan@gmail.com" }]
keywords = ["reinforcement-learning", "deep-learning", "ppo", "sac", "ddpg", "gymnasium", "isaaclab", "ros2"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "tensorboard>=2.13.0",
    "pyyaml>=6.0",
    "safetensors>=0.4.0",
]

[project.urls]
Homepage      = "https://github.com/Bigkatoan/SRL"
Documentation = "https://bigkatoan.github.io/SRL"
Repository    = "https://github.com/Bigkatoan/SRL"
Issues        = "https://github.com/Bigkatoan/SRL/issues"
Changelog     = "https://github.com/Bigkatoan/SRL/releases"

[project.optional-dependencies]
mujoco = ["gymnasium[mujoco]>=0.29.0", "mujoco>=3.0.0"]
box2d = ["box2d>=2.3.10", "pygame>=2.5.0"]
robotics = ["gymnasium-robotics>=1.2.0"]
racecar = [
    "racecar-gym @ git+https://github.com/axelbr/racecar_gym.git ; python_version < '3.11'",
    "yamldataclassconfig==1.5.0 ; python_version < '3.11'",
    "pybullet>=3.2.5",
]
all = [
    "gymnasium[mujoco]>=0.29.0",
    "mujoco>=3.0.0",
    "box2d>=2.3.10",
    "pygame>=2.5.0",
    "gymnasium-robotics>=1.2.0",
    "pybullet>=3.2.5",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
srl-train = "srl.cli.train:main"
srl-benchmark = "srl.cli.benchmark:main"
srl-visualize = "srl.cli.visualize:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["srl*"]

[tool.setuptools.package-data]
"srl" = ["py.typed"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
exclude = ["build", "matrix_runs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
